Implemented Attack training.
The attack training is in the new train file, the old blockbot is legacy
now. That being said, block has not been moved over to te new file.
There were various optimizations added, and it runs much smoother
with attack than block.

I may eventually move block over, but for the moment it is fine where
it is. The task of moving it over involves a bit of translating to make
it work with the new system. Also, moving the mouse causes quite a
bit of lag; so I am unsure of how much benefit it wold get by moving.